fix doctests, add procesopts to eventlisteners
Fix doctests so that they pin supervisor and superlance.

Added procesoptions to the syntax in the recipe for eventlisteners.
According to supervisor documentation eventlisteners are a special case
of progams and I have a use case where I want to start a listener after
x seconds, which needs startsecs=x.