Ensure buildout has information about dependencies on every run.
For this we need to call the installation of scripts and extra eggs on
every run, i.e.: during update as well as install. Without this,
supervisor and the eggs listed in the plugins option will appear and
disappear from the picked eggs list, e.g.: they appear on first run,
but are removed on subsequent runs, unless you manually trigger
another install of the part.